Fixed SP shuffling, SP order switch selection
SP shuffling algorithms (as well as tests) was flawed, always evaluated to true.

Previously SP order could be switched with any order 1-NumberOfStoryPieces. After validating the requirements, this selection should be 1-MaxUsedOrder where MaxUsedOrder is the maximum order that has ever been given to a StoryPiece on creation. This specification is now fulfilled.